pyproject.toml
[tool.poetry]
name = "pureshinshi"
version = "5.0.0"
description = "Shinshi is a modern and user-friendly Discord bot designed to give you and your servers great features and stable performance."
authors = ["stefanlight <stefanlight@vk.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
hikari = "^2.0.0.dev121"
hikari-crescent = "^0.6.3"
hikari-miru = "^3.2.2"
psutil = "^5.9.5"
humanize = "^4.8.0"
pyyaml = "^6.0.1"
uvloop = "^0.17.0"
python-dotenv = "^1.0.0"
floodgate-rs = "^0.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
